Moving the sharing pill in the system layer.
