+++ extensions/amp-story/1.0/amp-story-desktop.css
@@ -295,166 +295,3 @@ amp-story[standalone][desktop] {
   transform: translateX(-8px)!important;
   opacity: 1!important;
 }
-
-/* Top bar */
-div.i-amphtml-story-top {
-  position: relative!important;
-  display: block!important;
-  height: 96px!important;
-  width: 100%!important;
-  z-index: 1!important;
-  background: linear-gradient(to bottom, rgba(33,33,33,0.32) 0%, rgba(33,33,33,0) 100%)!important;
-}
-
-.i-amphtml-story-share-pill {
-  display: none !important;
-  position: absolute!important;
-  top: 0!important;
-  right: 40px!important;
-  bottom: 0!important;
-  margin: auto!important;
-  height: 40px!important;
-  align-items: center!important;
-}
-
-[desktop] .i-amphtml-story-share-pill {
-  display: flex !important;
-}
-
-/* Share text for share box */
-span.i-amphtml-story-share-pill-label {
-  font-family: 'Roboto', sans-serif!important;
-  position: absolute!important;
-  right: 15px!important;
-  text-align: center!important;
-  width: 60px!important;
-  font-size: 12px!important;
-  font-weight: bold!important;
-  letter-spacing: .5px!important;
-  margin: auto!important;
-  color:#fff!important;
-  box-sizing: initial !important;
-  text-transform: uppercase !important;
-}
-
-/* background for the share box */
-.i-amphtml-story-share-pill:before {
-  position: absolute!important;
-  content: ''!important;
-  width: 58px!important;
-  height: 40px!important;
-  padding: 0 16px!important;
-  background-color: rgba(255, 255, 255, .2)!important;
-  border-radius: 500px!important;
-  box-sizing: initial !important;
-  transition: width 300ms ease-in-out!important;
-  right: 0!important;
-}
-
-.i-amphtml-story-share-pill .i-amphtml-story-share-item  {
-  opacity: 0!important;
-  transition: opacity 300ms linear!important;
-}
-
-.i-amphtml-story-share-pill:hover .i-amphtml-story-share-item {
-  opacity: 1!important;
-}
-
-.i-amphtml-story-share-pill .i-amphtml-story-share-item {
-  margin: 0 !important;
-  width: 48px !important;
-  height: 48px !important;
-  padding: 0 !important;
-}
-
-/* Incremental fade in */
-.i-amphtml-story-share-pill:hover
-    .i-amphtml-story-share-item:nth-child(9) {
-  transition-delay: calc(30ms * 1)!important;
-}
-.i-amphtml-story-share-pill:hover
-    .i-amphtml-story-share-item:nth-child(8) {
-  transition-delay: calc(30ms * 2)!important;
-}
-.i-amphtml-story-share-pill:hover
-    .i-amphtml-story-share-item:nth-child(7) {
-  transition-delay: calc(30ms * 3)!important;
-}
-.i-amphtml-story-share-pill:hover
-    .i-amphtml-story-share-item:nth-child(6) {
-  transition-delay: calc(30ms * 4)!important;
-}
-.i-amphtml-story-share-pill:hover
-    .i-amphtml-story-share-item:nth-child(5) {
-  transition-delay: calc(30ms * 5)!important;
-}
-.i-amphtml-story-share-pill:hover
-    .i-amphtml-story-share-item:nth-child(4) {
-  transition-delay: calc(30ms * 6)!important;
-}
-.i-amphtml-story-share-pill:hover
-    .i-amphtml-story-share-item:nth-child(3) {
-  transition-delay: calc(30ms * 7)!important;
-}
-.i-amphtml-story-share-pill:hover
-    .i-amphtml-story-share-item:nth-child(2) {
-  transition-delay: calc(30ms * 8)!important;
-}
-.i-amphtml-story-share-pill:hover
-    .i-amphtml-story-share-item:nth-child(1) {
-  transition-delay: calc(30ms * 9)!important;
-}
-
-.i-amphtml-story-share-pill .i-amphtml-story-share-list {
-  padding: 0!important;
-  margin: 0 72px 0 0 !important;
-}
-
-.i-amphtml-story-share-pill:hover:before {
-  width: calc(100% - 32px) !important;
-}
-
-.i-amphtml-story-share-pill .i-amphtml-story-share-icon {
-  display: flex !important;
-  justify-content: center;
-  align-items: center;
-  width: 48px !important;
-  height: 48px !important;
-  padding: 0;
-}
-
-.i-amphtml-story-ui-no-buttons .i-amphtml-story-share-pill {
-  visibility: hidden !important;
-}
-
-[desktop] .i-amphtml-story-share-icon {
-  font-size: 0px!important;
-  background-size: 24px 24px !important;
-  background-position: center center !important;
-  padding: 0 !important;
-}
-
-[desktop] .i-amphtml-story-share-icon.amp-social-share-facebook {
-  /* SVG icon has inner padding, so the glyph renders visually smaller than
-   * other icons, so we work around it by setting a higher background size. */
-  background-size: 28px 28px !important;
-  background-position: center center !important;
-}
-
-[desktop] .i-amphtml-story-share-icon::before {
-  display: none !important;
-}
-
-[desktop] .i-amphtml-story-share-icon:active::before {
-  display: none !important;
-}
-
-[desktop] .i-amphtml-story-share-list {
-  margin-right: 72px !important;
-}
-
-/* system layer changes */
-
-[desktop] .i-amphtml-story-top {
-  z-index: 100002 !important; /* 1 higher than bookend */
-}
