+++ extensions/amp-story/1.0/amp-story-system-layer.js
@@ -16,14 +16,16 @@
 import {Action, StateProperty} from './amp-story-store-service';
 import {CSS} from '../../../build/amp-story-system-layer-1.0.css';
 import {DevelopmentModeLog, DevelopmentModeLogButtonSet} from './development-ui';
+import {LocalizedStringId} from './localization';
 import {ProgressBar} from './progress-bar';
 import {Services} from '../../../src/services';
+import {ShareWidget} from './amp-story-share';
 import {createShadowRootWithStyle} from './utils';
 import {dev} from '../../../src/log';
 import {dict} from '../../../src/utils/object';
 import {getMode} from '../../../src/mode';
 import {matches} from '../../../src/dom';
-import {renderAsElement} from './simple-template';
+import {renderAsElement, renderSimpleTemplate} from './simple-template';
 
 
 /** @private @const {string} */
@@ -79,6 +81,30 @@ const TEMPLATE = {
 
 
 /**
+ * Container for "pill-style" share widget, rendered on desktop.
+ * @private @const {!./simple-template.ElementDef}
+ */
+const SHARE_WIDGET_PILL_CONTAINER = {
+  tag: 'div',
+  attrs: dict({'class': 'i-amphtml-story-share-pill-container'}),
+  children: [
+    {
+      tag: 'div',
+      attrs: dict({'class': 'i-amphtml-story-share-pill'}),
+      children: [
+        {
+          tag: 'span',
+          attrs: dict({'class': 'i-amphtml-story-share-pill-label'}),
+          localizedStringId:
+              LocalizedStringId.AMP_STORY_SYSTEM_LAYER_SHARE_WIDGET_LABEL,
+        },
+      ],
+    },
+  ],
+};
+
+
+/**
  * System Layer (i.e. UI Chrome) for <amp-story>.
  * Chrome contains:
  *   - mute/unmute button
@@ -89,9 +115,12 @@ export class SystemLayer {
   /**
    * @param {!Window} win
    */
-  constructor(win) {
+  constructor(win, ampdoc) {
     /** @private {!Window} */
     this.win_ = win;
+
+    /** @private {!../../../src/service/ampdoc-impl.AmpDoc} */
+    this.ampdoc_ = ampdoc;
 
     /** @private {boolean} */
     this.isBuilt_ = false;
@@ -119,6 +148,9 @@ export class SystemLayer {
 
     /** @private {!DevelopmentModeLogButtonSet} */
     this.developerButtons_ = DevelopmentModeLogButtonSet.create(win);
+
+    /** @private {?Node} */
+    this.sharePillContainerNode_ = null;
 
     /** @private @const {!./amp-story-store-service.AmpStoryStoreService} */
     this.storeService_ = Services.storyStoreService(this.win_);
@@ -261,6 +293,10 @@ export class SystemLayer {
    * @private
    */
   onDesktopStateUpdate_(isDesktop) {
+    if (isDesktop) {
+      this.buildSharePill_();
+    }
+
     this.vsync_.mutate(() => {
       isDesktop ?
         this.getShadowRoot().setAttribute('desktop', '') :
@@ -336,6 +372,27 @@ export class SystemLayer {
   updateProgress(pageId, progress) {
     // TODO(newmuis) avoid passing progress logic through system-layer
     this.progressBar_.updateProgress(pageId, progress);
+  }
+
+  /**
+   * Builds and appends the share pill. Desktop only.
+   * @private
+   */
+  buildSharePill_() {
+    if (this.sharePillContainerNode_) {
+      return;
+    }
+
+    this.sharePillContainerNode_ =
+        renderSimpleTemplate(this.win_.document, SHARE_WIDGET_PILL_CONTAINER);
+
+    const shareWidget = new ShareWidget(this.win_);
+
+    this.sharePillContainerNode_
+        .querySelector('.i-amphtml-story-share-pill')
+        .appendChild(shareWidget.build(this.ampdoc_));
+
+    this.systemLayerEl_.appendChild(this.sharePillContainerNode_);
   }
 
   /**
